real-time-analytics-platform/pipeline/requirements.txt
apache-airflow==2.7.0
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9